+++ HISTORY.md
@@ -42,6 +42,7 @@
 
 #### Improvements
 
+- Remove operators from GaussianEmitter and IsoLineEmitter ({pr}`508`)
 - Update QDax visualizations to match QDax 0.5.0 ({pr}`502`)
 - Skip qdax tests if qdax not installed ({pr}`491`)
 - Move yapf after isort in pre-commit ({pr}`490`)